requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.3.0
billiard==4.2.1
python-docx==1.1.2
bitarray==2.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
grpcio==1.71.0
portalocker==3.1.1
contourpy==1.2.0
cssselect==1.2.0
curl-cffi==0.7.4
cycler==0.12.1
cytoolz==0.12.2
dataclasses-json==0.6.7
datarecorder==3.6.2
deprecated==1.2.15
distro==1.9.0
dnspython==2.5.0
downloadkit==2.0.7
drissionpage==4.0.5.6
ecdsa==0.19.0
email-validator==2.1.0.post1
et-xmlfile==2.0.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.2.2
fake-useragent==1.5.1
fastapi==0.101.1
fasteners==0.19
filelock==3.17.0
fonttools==4.46.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.1.2
jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kafka-python==2.0.2
kiwisolver==1.4.5
kombu==5.4.2
lru-dict==1.2.0
lxml==5.3.0
mako==1.3.9
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
motor==3.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.61.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.9.0
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
prompt-toolkit==3.0.49
propcache==0.2.1
protobuf==4.25.1
psutil==7.0.0
pyasn1==0.6.0
pychrome==0.2.4
pycparser==2.22
pycryptodome==3.19.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-extra-types==2.5.0
pydantic-settings==2.5.2
pygments==2.19.1
pyjwt==1.7.1
aiokafka==0.10.0
aiosmtplib==3.0.1
pymongo==4.4.1
pyparsing==3.1.1
pytelegrambotapi==4.22.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyunormalize==15.1.0
pyvirtualdisplay==3.0
pyyaml==6.0.2
redis==5.2.1
referencing==0.32.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rlp==4.0.0
rpds-py==0.15.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.38
standard-imghdr==3.13.0
starlette==0.27.0
stripe==11.4.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.0
toolz==0.12.0
tqdm==4.67.1
ttkbootstrap==1.10.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
ujson==5.9.0
urllib3==2.3.0
uvicorn==0.23.2
uvloop==0.19.0
vine==5.1.0
watchfiles==0.19.0
wcwidth==0.2.13
web3==6.12.0
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0